Feat: Add books page with CRUD functionality for managing books
